backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai>=1.10.0
pinecone-client>=3.0.0
beautifulsoup4==4.12.3
requests==2.31.0
python-multipart==0.0.6
pypdf==4.0.1
python-dotenv==1.0.1
urllib3==2.2.0